[Chore] modify profile page 코드 위치 수정
